requirements.txt
requests
pyfy
python-vlc
pafy
